@@ -4,20 +4,19 @@
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
-    "@testing-library/user-event": "^14.4.3",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
     "@testing-library/dom": "^8.17.1",
-    "@types/jest": "^28.1.6",
-    "@types/node": "^16.11.47",
-    "@types/react": "^17.0.48",
-    "@types/react-dom": "^17.0.17",
-    "@types/react-router-dom": "^5.3.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^18.11.15",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-guard": "link:..",
-    "react-router-dom": "^5.3.3",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.1.2",
+    "react-router-dom": "^6.4.5",
+    "react-scripts": "^5.0.1",
+    "typescript": "^4.9.4",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
