@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
@@ -9,6 +8,7 @@ import reactHookGuard from 'react-hook-guard';
 import Spin from './components/Spin';
 import AccessDenied from './components/AccessDenied';
 import NotFound from './components/NotFound';
+import {createRoot} from "react-dom/client";
 
 reactHookGuard.config({
     SuspenseFallback: Spin,
@@ -17,9 +17,11 @@ reactHookGuard.config({
         path: '',
         component: NotFound,
     }
-})
+});
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root')!);
+
+root.render(
   <React.StrictMode>
       <BrowserRouter>
           <AuthProvider>
@@ -27,7 +29,6 @@ ReactDOM.render(
           </AuthProvider>
       </BrowserRouter>
   </React.StrictMode>,
-  document.getElementById('root')
 );
 
 // If you want to start measuring performance in your app, pass a function
